add foreigh keys check for soft delete

+++ internal/backend/db/dbgorm/triggers.go
@@ -29,6 +29,14 @@ func (gdb *gormdb) deleteTrigger(ctx context.Context, triggerID sdktypes.UUID) e
 		if err := tx.isCtxUserEntity(tx.ctx, triggerID); err != nil {
 			return err
 		}
+
+		// enforce foreign keys constrains while soft-deleting
+		var count int64
+		tx.db.Model(&scheme.Event{}).Where("deleted_at is NULL and trigger_id = ?", triggerID).Count(&count)
+		if count > 0 {
+			return fmt.Errorf("FOREIGN KEY: %w", gorm.ErrForeignKeyViolated)
+		}
+
 		return tx.db.Delete(&scheme.Trigger{TriggerID: triggerID}).Error
 	})
 }
